# feat: optimize `CachedRequestBuilder`