--- a/Refit/CachedRequestBuilderImplementation.cs
+++ b/Refit/CachedRequestBuilderImplementation.cs
@@ -20,10 +20,10 @@ namespace Refit
         }
 
         readonly IRequestBuilder innerBuilder;
-        readonly ConcurrentDictionary<
-            string,
+        internal readonly ConcurrentDictionary<
+            MethodTableKey,
             Func<HttpClient, object[], object?>
-        > methodDictionary = new();
+        > MethodDictionary = new();
 
         public Func<HttpClient, object[], object?> BuildRestResultFuncForMethod(
             string methodName,
@@ -31,13 +31,22 @@ namespace Refit
             Type[]? genericArgumentTypes = null
         )
         {
-            var cacheKey = GetCacheKey(
+            var cacheKey = new MethodTableKey(
                 methodName,
                 parameterTypes ?? Array.Empty<Type>(),
                 genericArgumentTypes ?? Array.Empty<Type>()
             );
-            var func = methodDictionary.GetOrAdd(
-                cacheKey,
+
+            if (MethodDictionary.TryGetValue(cacheKey, out var methodFunc))
+            {
+                return methodFunc;
+            }
+
+            // use GetOrAdd with cloned array method table key. This prevents the array from being modified, breaking the dictionary.
+            var func = MethodDictionary.GetOrAdd(
+                new MethodTableKey(methodName,
+                    parameterTypes?.ToArray() ?? Array.Empty<Type>(),
+                    genericArgumentTypes?.ToArray() ?? Array.Empty<Type>()),
                 _ =>
                     innerBuilder.BuildRestResultFuncForMethod(
                         methodName,
@@ -48,37 +57,88 @@ namespace Refit
 
             return func;
         }
+    }
 
-        static string GetCacheKey(
-            string methodName,
-            Type[] parameterTypes,
-            Type[] genericArgumentTypes
-        )
+    /// <summary>
+    /// Represents a method composed of its name, generic arguments and parameters.
+    /// </summary>
+    internal readonly struct MethodTableKey : IEquatable<MethodTableKey>
+    {
+        /// <summary>
+        /// Constructs an instance of <see cref="MethodTableKey"/>.
+        /// </summary>
+        /// <param name="methodName">Represents the methods name.</param>
+        /// <param name="parameters">Array containing the methods parameters.</param>
+        /// <param name="genericArguments">Array containing the methods generic arguments.</param>
+        public MethodTableKey (string methodName, Type[] parameters, Type[] genericArguments)
         {
-            var genericDefinition = GetGenericString(genericArgumentTypes);
-            var argumentString = GetArgumentString(parameterTypes);
-
-            return $"{methodName}{genericDefinition}({argumentString})";
+            MethodName = methodName;
+            Parameters = parameters;
+            GenericArguments = genericArguments;
         }
 
-        static string GetArgumentString(Type[] parameterTypes)
+        /// <summary>
+        /// The methods name.
+        /// </summary>
+        string MethodName { get; }
+
+        /// <summary>
+        /// Array containing the methods parameters.
+        /// </summary>
+        Type[] Parameters { get; }
+
+        /// <summary>
+        /// Array containing the methods generic arguments.
+        /// </summary>
+        Type[] GenericArguments { get; }
+
+        public override int GetHashCode()
         {
-            if (parameterTypes == null || parameterTypes.Length == 0)
+            unchecked
             {
-                return "";
+                var hashCode = MethodName.GetHashCode();
+
+                foreach (var argument in Parameters)
+                {
+                    hashCode = (hashCode * 397) ^ argument.GetHashCode();
+                }
+
+                foreach (var genericArgument in GenericArguments)
+                {
+                    hashCode = (hashCode * 397) ^ genericArgument.GetHashCode();
+                }
+                return hashCode;
+            }
+        }
+
+        public bool Equals(MethodTableKey other)
+        {
+            if (Parameters.Length != other.Parameters.Length
+                || GenericArguments.Length != other.GenericArguments.Length
+                || MethodName != other.MethodName)
+            {
+                return false;
             }
 
-            return string.Join(", ", parameterTypes.Select(t => t.FullName));
-        }
-
-        static string GetGenericString(Type[] genericArgumentTypes)
-        {
-            if (genericArgumentTypes == null || genericArgumentTypes.Length == 0)
+            for (var i = 0; i < Parameters.Length; i++)
             {
-                return "";
+                if (Parameters[i] != other.Parameters[i])
+                {
+                    return false;
+                }
             }
 
-            return "<" + string.Join(", ", genericArgumentTypes.Select(t => t.FullName)) + ">";
+            for (var i = 0; i < GenericArguments.Length; i++)
+            {
+                if (GenericArguments[i] != other.GenericArguments[i])
+                {
+                    return false;
+                }
+            }
+
+            return true;
         }
+
+        public override bool Equals(object? obj) => obj is MethodTableKey other && Equals(other);
     }
 }