Rewrote `CachedRequestBuilder` to use a custom `MethodTableKey` value as a dictionary key instead of a concatenated string. Resulted in some modest memory savings and reduced object usage.

- Added unit tests to ensure that the dictionary key works correctly.
- On lines 48 & 49 of `CachedRequestBuilderImplementation` I copy the arrays to create the dictionary key, this is to prevent the arrays being changed, breaking the dictionary. This can be removed if Refit assumes that the input arrays are not accessible to the user.
- Implemented `IEquatable<MethodTableKey>`, implementing my own `GetHashCode` method. I could have used the recommended `HashCode` object but it is not available in .NET Framework. I could copy the implementation over and use it if you want.

### Original
| Method             | Mean     | Error    | StdDev   | Gen0   | Gen1   | Allocated |
|------------------- |---------:|---------:|---------:|-------:|-------:|----------:|
| TaskT_Async        | 16.64 us | 0.329 us | 0.885 us | 1.4038 | 0.0305 |  13.04 KB |
| TaskTLong_Async    | 17.25 us | 0.343 us | 0.669 us | 1.4954 | 0.0305 |  13.82 KB |
| TaskTGeneric_Async | 18.03 us | 0.360 us | 0.592 us | 1.6479 | 0.0305 |  15.19 KB |

## With changes
| Method             | Mean     | Error    | StdDev   | Gen0   | Gen1   | Allocated |
|------------------- |---------:|---------:|---------:|-------:|-------:|----------:|
| TaskT_Async        | 12.19 us | 0.241 us | 0.277 us | 1.4038 | 0.0305 |  12.92 KB |
| TaskTLong_Async    | 13.26 us | 0.262 us | 0.331 us | 1.4648 | 0.0305 |  13.51 KB |
| TaskTGeneric_Async | 14.35 us | 0.286 us | 0.627 us | 1.5564 | 0.0305 |  14.49 KB |